Move linters into their own config
